+++ frontend/memo-scholar/src/hooks/useTheme.ts
@@ -1,0 +1,9 @@
+import { useEffect, useState } from "react";
+
+export function useTheme() {
+  const [isDark, setIsDark] = useState(false);
+  useEffect(() => {
+    document.documentElement.classList.toggle("dark", isDark);
+  }, [isDark]);
+  return { isDark, toggle: () => setIsDark(d => !d) };
+}
